SafeEntryGuard/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safeentryguard"
version = "0.1.0"
description = "SafeEntryGuard: a practical post-generation filter for selecting the safest and most direct website from LLM answers."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "dnspython>=2.6.1",
  "requests>=2.31.0",
  "tldextract>=5.1.2",
]

[project.scripts]
safeentryguard = "safeentryguard.cli:main_entry"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
